requirements.txt
altgraph==0.17.3
auto-py-to-exe==2.30.0
bottle==0.12.23
bottle-websocket==0.2.9
cffi==1.15.1
Eel==0.14.0
flake8==5.0.4
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
importlib-metadata==4.2.0
mccabe==0.7.0
numpy
pefile==2023.2.7
Pillow==9.4.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pymodbus==2.5.3
pyparsing==3.0.9
PyQt5==5.15.8
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyserial==3.5
pywin32==305
pywin32-ctypes==0.2.0
six==1.16.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
whichcraft==0.6.1
zipp==3.12.0
zope.event==4.6
zope.interface==5.5.2
openpyxl==3.1.2
